requirements.txt
flask==3.0.3
flask-cors==4.0.1
opencv-python==4.10.0.84
numpy==1.24.3
faiss-cpu==1.8.0
mediapipe==0.10.14
psutil==5.9.8
